+++ src/main/resources/templates/mypage-edit.html
@@ -14,7 +14,7 @@
             </div>
         </div>
     </div>
-<form th:method="put" action="/accounts/profile/edit" th:object="${userForm}">
+<form th:method="put" action="/accounts/profile/edit" th:object="${userDto}">
     <div class="container-fluid max-width-lg mx-auto mrg-btm-30 padding-30">
         <div class="row">
             <div class="col-md-7 mx-auto">
